feat: separate modal library component
